ctdb-common: Fix signed/unsigned comparisons by casting
In one case, given triviality of change, add missing braces and fix
whitespace.

+++ ctdb/common/path.c
@@ -89,7 +89,7 @@ static bool path_construct(char *path, const char *subdir)
 			       subdir);
 	}
 
-	if (len >= sizeof(p)) {
+	if ((size_t)len >= sizeof(p)) {
 		return false;
 	}
